Expand Room database caching and improve search UI robustness.
### Core Storage & Persistence
- **Implement Series Caching**: Add comprehensive Room entities and DAOs for TV series, including `SeriesEntity`, `SeasonEntity`, `CreatedByEntity`, `NetworksEntity`, and `LastEpisodeEntity`.
- **Enhanced Movie Caching**: Add `BelongsToCollectionEntity` to store movie collection metadata.
- **Refactored Cache Logic**:
    - Relocate `CacheDao` and `CacheRepositoryImpl` to a dedicated `:room:details` package.
    - Implement a `Cacheable` interface with `isDataValid` logic to handle cache expiration (set to 24 hours).
    - Introduce cross-reference tables for series genres, production companies, countries, languages, and videos.
    - Update `CacheRepository` to support conditional fetching based on `appendToResponse` parameters (credits, videos).
- **Database Schema**: Update `AppDatabase` to version 1 with new entities and implement `Converters` for handling `List<Int>` and `List<String>` via Kotlin Serialization.

### Network & Data
- **Resilient API Calls**: Introduce a `retry` mechanism in the network layer with exponential backoff for TMDB API requests.
- **Improved Paging**: Add null-safety checks to `BasePagingSource` and `TimedPagingSource` to prevent crashes on empty responses.
- **Refactored Result Handling**: Rename `safeApiCall` to `toResult` and implement `coRunCatching` for safer coroutine execution.
- **Data Mapping**: Support `BelongsToCollection` in movie DTOs and domain models; update `Network` and `Job` models to use `Int` for IDs and counts.

### UI & Features
- **Search Enhancements**:
    - Add a minimum character threshold (2) to `SearchViewModel` before triggering API calls.
    - Implement `FocusManager.clearFocus()` in `SearchScreen` to dismiss the keyboard when scrolling results.
    - Update `SearchScreen` to better handle `Waiting` and `Error` states with proper padding and background styling.
- **Detail Screen Improvements**:
    - Refine `FullCreditsScreen` to display roles for cast and job/department combinations for crew members.
    - Update `MovieViewModel` to dynamically switch between Movie `credits` and Series `aggregateCredits`.
- **Navigation & Cleanup**: Reorganize repository implementations into consistent package structures (e.g., moving `RecentSearch` logic to `:room:search`).

+++ feature/tmdb/src/main/java/com/vroff/tmdb/entity/common/JobDTO.kt
@@ -8,7 +8,7 @@ data class JobDTO(
     val creditId: String,
     val job: String,
     @SerializedName("episode_count")
-    val episodeCount: Long,
+    val episodeCount: Int,
 ) {
     fun mapToDomain(): Job =
         Job(
